Force the use of an intermediate PRInitCreator object for constructing ProjectedRowInitializer. This costs an extra O(|col_ids|) lookup per projected row initialized. I don't know if the code reuse is worth the cost, suggestions are welcome. Names are hard, if we stick with this approach I'll try to come up with better names.
To recap: we would like to reuse ProjectedRowInitializer's sorting/padding/caching logic for constructing ProjectedRow. This is for indexes: we will have a type[] which we convert to attr_sizes[] which we want to turn into a ProjectedRow. But we don't want to use attr_sizes to construct a BlockLayout; that creates NUM_RESERVED_COLUMNS of other stuff and tracks its own unnecessary metadata. So we're trying to create and expose alternative APIs for ProjectedRowInitializer.