Merge of split implementations (#1)
Signed off on during team meeting (2019-03-21T17:00):

* Finish refactor to std::map

Implementation is correct for the sequential case but does not actually support
any versioning.  It is simply a port of the existing code to point into the map
to ensure compilation and unittest success before further changes.

* Remove sql_table_concurrent_test

The test looks at behavior under invalid circumstances.  Catalog
will prevent multiple in-flight schema changes by forcing an abort
when write-write conflict occurs on the table's schema version
in the catalog.