halt the macos workers after the build
the new version of the supervisor script will reinstantiate the worker
between each build.  this should resolve the problems with the macos
workers getting corrupted working repo on git checkout when they
(presumably) run out of disk space.

on the downside, things may be in the queue a little longer while the
VMs stop and start.