Tpcc benchmark log (#701)
Added the ability to correctly control the # of threads used in benchmarks. We also now use BenchmarkConfig::logfile_path to control where we write the WAL during benchmarks.
I also modified the behavior of WorkerPool so that the constructor does not automatically call StartUp(). This also includes the delete fix from @rohanaggarwal7997 in #690. Props to @thepinetree for staying true to the cause.