Infer NULL type during bytecode generation.
Address Prashanth's comments on Friday. We do not want to perform
NULL type inference during bytecode interpretation as that would
be unneccessary runtime overhead. Currently, all SQL types can
rely on the Val.is_null field for null checking, and so it is
sufficient to add a Val::Null() constructor and to call that
directly in OpInitSqlNull for most of the types. In the future,
if new SQL types need custom NULL behavior, the switching on the
type should also occur during bytecode generation.