Add support for NULL. (#768)
* Add support for NULL and fix null columns not advancing in packet writer.

* Address half of Prashanth's comments. For the other half: he prefers Null(&type) to having a NullType() per type, because it reduces the number of builtins overall, but we'll roll with this for now and change it later if necessary.

* Address Prashanth's comments, round two.

* Address review comments; rewrite null initialization to be NullToSql(&expr_with_desired_type) instead of one builtin per type.

* Add TODO comment for changing Bool->Boolean in codegen.

* Fix broken is_null_predicate_test.

* UNREACHABLE() instead of TERRIER_ASSERT() so that it doesn't bomb out in release.

* Fix trying to turn subqueries into nulls during binding.

* Infer NULL type during bytecode generation.

Address Prashanth's comments on Friday. We do not want to perform
NULL type inference during bytecode interpretation as that would
be unneccessary runtime overhead. Currently, all SQL types can
rely on the Val.is_null field for null checking, and so it is
sufficient to add a Val::Null() constructor and to call that
directly in OpInitSqlNull for most of the types. In the future,
if new SQL types need custom NULL behavior, the switching on the
type should also occur during bytecode generation.

* Take in @nullToSql(&e) instead of @nullToSql(e).

Fixes bug with InitSqlNull's opcode taking too many args,
which resulted in strange errors during JIT.

Fixes bug with @isSqlNull() not taking lvalue.

Co-authored-by: Matt Butrovich <mbutrovich@users.noreply.github.com>