Stop inheritance hacks for AnalyzeBottomTranslator
It turns out that making AnalyzeBottomTranslator inherit directly from
StaticAggregateBottomTranslator doesn't really work, since apparently
that brain stuff actually matters (or maybe just that GetOp() returns
the right thing).

So, photocopying the class. This is unfortunate but I don't know if
there is really a better alternative at this time.