Redo TopKAggregate to behave like other aggregates
The purpose of these changes is to allow the TopKAggregate function to
be treated similarly to other aggregates, thereby making it easier to
work with in the Analyze code.

The previous design used HasResult and Result functions, and required
using entirely separate builtins. There were a few minor issues with
this:

* The naming is a bit unfortunate, since it implies a similarity to the
  other aggregators' interfaces that does not exist. Since the interface
  is more similar to an iterator, the functions should be named HasNext
  and Next respectively.
* There is an efficiency issue: Result required generating the top-K
  elements vector from the underlying TopKElements object on every call,
  which is essentially quadratic in K.

I am not sure what guided the design of the previous design: if there is
a need for an iterator-like interface from TPL, then it would be good to
restore it after resolving the above two issues. But I have removed it
for now.

For my current purposes, I have changed the TopKAggregate to output a
serialized binary string when returning a result to SQL. Since this has
the same interface (modulo types) as other aggregates, this makes it
possible to use with existing code such as AggregateHelper.

As such, I also implemented a serialize() function for the
madoka::Sketch class, and I removed special code for the TopKAggregate
builtins, preferring to share code with existing aggregates instead.

I also added support for TOP_K_AGGREGATE to be used with
AggregateExpression, in preparation for its impending use within
AnalyzeTranslator.

TODO: code to test that TopKAggregate serialization works, and code to
implement deserialization.

TODO: update or delete topk.tpl to reflect the changes to the
TopKAggregate interface and builtins.