Use regions to allocate plan generator expressions
Currently, a large number of allocations are performed as part of physical plan generation.
This presents a number of issues:
1. A large number of small allocations in a hottish path adds unnecessary overhead.
2. Managing the newly allocated memory adds unnecessary code complexity and further overhead.
3. Presenting a large number of expenive allocations encourages/normalizes expensive code in hottish paths.

We can address the performance and complexity issues by creating a per-connection-context region allocator.

This addresses 1. by using a low-overhead allocator.
This addresses 2. by deferring frees to the closure of a connection context.
This addresses 3. by moving to a quicker routine.

It may be possible to avoid allocations entirely in certain cases going into the future.