Manually deregister exception frames
LLVM can "forget" to properly cleanup its exception frames in certain cases when dealing with ELFs. An important such case is when an exception is thrown. We throw many exceptions, so this hurts us in an especially rough way. That's because when LLVM forgets to clean up its exception frames, exceptions thrown later are undefined behavior. (UAF in libunwind)
Issue #961 (https://github.com/cmu-db/terrier/issues/961) is such an instance of this undefined behavior causing a crash.

Hopefully fixes #961 by forcing LLVM to clear EH frames when the CompiledModule destructor is called.
Note that (as long as we don't need the EH frames anymore) it is safe to call deregisterEHFrames even if it has been called before. This is because there's an LLVM-internal vectors that tracks registered EH frames. If you call it more than once, it will be empty on the next-call, making it essentially a no-op.

When LLVM's JITLink project is finished and we move to an LLVM version that supports it, it would be best to migrate to it from RuntimeDyld, which has serious issues with EH frames. That would, in theory, fix this issue entirely.
I have no idea if this solution is thread safe or not, but LLVM doesn't lock on it and it should only be called rarely, so I think it is. __deregister_frame has...interesting documentation, to say the least: https://gcc.gnu.org/onlinedocs/gcc-4.2.4/gccint/Exception-handling-routines.html