Build improvements (#1026)
* Add support for compiling shared library

* mac OS build fix

Searches for Brew's clang++ on mac OS, which is required for builds.

* Build with dynamic-linking on non-mac OS platforms

Supports dynamic linking on non-mac OS platforms, but force-disables it on mac OS.

* Make sure that OLTP-Bench results directory names are unique (#1015)

* Move Limit/Offset Fields to AbstractScanPlanNode (#938)

Co-authored-by: Matt Butrovich <mbutrovich@users.noreply.github.com>
Co-authored-by: Andy Pavlo <pavlo@cs.brown.edu>

* Merge Matt's requested changes

Implements the changes that Matt requested.

* Disable ODR violation detection for asan tests

Address sanitizer checks for one-definition-rule violations. This ruins tests.
This is technically a bandaid solution, but since the violations don't seem to matter, it's probably the best solution.

* Fix ASAN options

Avoids overwriting Mac OS ASAN options.

* Statically link spdlog

An attempt (here goes nothing) at bringing in spdlog as a git submodule and statically linking it.

* Cleaner Clang initialization on MacOS

Sets the location of Clang on MacOS by reusing existing variables.

* Pull submodules from within Jenkinsfile

* Just a test commit to try to coerce a proper build. Will revert.

Ignore me.

* Revert "Explicit instantiation of spdlog template functions (#995)"

This reverts commit 686eb69f71220d4abeb3a916252f6ae6924927ba.

* Update spdlog entry in versions.txt

Migrate to remote fetching of spdlog.

* Delete spdlog folder

Entirely manage spdlog through CMake.

* Revert "Just a test commit to try to coerce a proper build. Will revert."

This reverts commit d162eaeb46307ad6a4f991004c9ff42da2e8a2e7.

* Revert "Pull submodules from within Jenkinsfile"

This reverts commit ca80090883b89833bd2900faae03038823cb5ba0.

* Revert changes to .gitmodules

* Revert src/CMakeLists.txt

* Fetch and (properly) build spdlog

* Literally just adding a file named "a"

Attempting to make this visible to Jenkins.

* Obscure bugfix for Release builds

Fixes Release builds for macOS by slightly changing header order.

May cause obscure AppleClang segfaults. In fact, this will almost certainly cause a segfault on the Jenkins machine during the build, but here's why I'm submitting this anyway:
I've been attempting to debug AppleClang, but after a few hours of debugging, I'm at my wits end.
There's a bug somewhere in this call to getRegClassWeight, which causes getRegClassWeight to dereference a null pointer: https://github.com/llvm/llvm-project-staging/blob/f0fd4f3af0acda5068acc5bbe10b3792c41bdfc4/llvm/lib/CodeGen/MachineLICM.cpp#L912
The really weird thing is that for regular Clang, getRegisterClassWeight should be essentially a no-op here--it's a const getter. There's really not much that should be going on unless a null registerclass is being passed, but even then, it shouldn't be crashing *here*. A null registerclass should cause catastrophic failure way earlier. So I'm thinking Apple's fork of Clang does something special here that Clang doesn't do, and that the Jenkins version of AppleClang might not incorporate this new feature, since it runs an older version of the macOS toolchain.

* Fix FindClangTools.cmake

A long-standing pet peeve of mine is that FindClangTools.cmake doesn't work (at least not on Arch) when clang-{format,tidy} are installed through the system package manager. This fixes that.

* Cleaner ThirdPartyToolchain.cmake

Removes some unneeded code that used to (but no longer) exists in the codebase.

* Suppress clang-format warning

We are actually required to keep the includes in this order (at least, spdlog.h MUST be at the top). clang-format is really bothered by this.

* Remove stray file. Override compiler on macOS in Jenkinsfile.

* Different way to set compiler on macOS.

* Revert "Different way to set compiler on macOS."

This reverts commit 8a06349e

* Fix typo.

Co-authored-by: Kevin Geng <khg@andrew.cmu.edu>
Co-authored-by: Jordi Gonzalez <jordig@imac1.db.pdl.local.cmu.edu>
Co-authored-by: Yao Yu <skyfish.jy@gmail.com>
Co-authored-by: Joseph Koshakow <jkosh44@users.noreply.github.com>
Co-authored-by: Matt Butrovich <mbutrovich@users.noreply.github.com>
Co-authored-by: Andy Pavlo <pavlo@cs.brown.edu>
Co-authored-by: Matt Butrovich <mbutrovich@gmail.com>