Miscellaneous test fix
Noise is apparently massive in some situations? Or I have really bad luck. In either case, we can make the signal *much* louder than the noise, and we can also reduce the amount of noise that exists in the first place. We do that in this commit by increasing the average number of samples taken (by running the filters 10x more) and increase the overhead of the "slow" filter.