Fix AdaptiveCheckTest in sql_filter_manager_test.cpp
Noise is apparently massive in some situations, which can cause spurious test failures. However, we can both decrease noise and increase the signal.
To do so, this commit increases the average number of samples taken and dramatically increasing the overhead of the "slow" filters, respectively.
For added effect, we exaggerate selectivity differences as well. This is mostly apparent in execution traces.
This should fix the test failures noted by Wan, allowing this test to be ran on Apple machines. It also fixes some spurious test failures that I've encountered.