Proper implementation of hash table iterator
Previous iterator did not work for tables built in parallel.

Also revert adding pipeline state to arguement for OuterLeftConsumer. The idea
is that we need to pass in the pipeline state as an arguement to prevent race
conditions, but currently I'm not sure if the pipeline state is accessible
from the specific section of the code. So, I'm going to not pass the pipeline
state until recieving further feedback.