Fix concat operator
The sql parser was correctly parsing "||" as CONCAT but there was
no tpl parsing, tpl ast generation, semantic checking, or
bytecode generation for concat. This commit adds tpl ast
generation, semantic checking, and bytecode generation. tpl
parsing still isn't working properly. We reuse the concat
built-in bytecode handling as the underlying implementation.

Originally CONCAT was being treated as a normal BinaryOpExpr,
however CONCAT needs access to the Execution Context in order
to allocate space for a new string. Normal BinaryOpExpr don't
have a reference to the execution context. Therefore we needed
to create a new type of expression, StringBinaryOpExpr, which
has a reference to the execution context.

StringBinaryOpExpr was made to be generalized for use with  any
string binary operation. Some other options would have been:
- Add a reference to the Execution Context to BinaryOpExpr.
However this is probably wasteful since the majority of binary
operations do not need a reference to the Execution Context.
- Create a specific operation for CONCAT.
- Generalize further and make some StringOpExpr, that instead
of having a left and right expression, has a vector of
expressions. This would allow it to be reused on string
operations that need a reference to the Execution Context and
have any amount of arguments.

Many elements were inspired by/taken out of the following PR:
https://github.com/cmu-db/terrier/pull/839/files

Fixes: #804

Co-authored-by: VenkatDatta <nhvenkatdatta@gmail.com>