Add support for variadic functions in the database catalog
When looking up a function in the pg_catalog.pg_proc, we the signature
from parsing the function, and namespace to lookup in a unique index.
This was an issue because variadic functions do not have a consistent
signature.

This commit changes the unique index to be non-unique and only use the
function name and namespace. We then compare the arg types returned
from the index to see if any would work with the parsed function.

Fixes #1176