Add messenger draft.
Messenger is responsible for the ZeroMQ related code.
MessengerLogic is responsible for acting on a received message and
interfacing with the rest of the system.
I suck at coming up with names, perhaps we can fix that.

Works with a Python client.
Current protocol: identifier | dummy | payload
Where payload's first byte indicates what callback should be invoked.
If you're familiar with ZeroMQ, this is just the REQ/REP protocol
made explicit so that we can use the async ROUTER/DEALER pattern.

Needs much more documentation so that people don't need to learn
ZeroMQ just to hack on the messenger. To people familiar with
ZeroMQ, this is a standard ROUTER-DEALER without a local inproc
worker pool on the server side. We may decide to add an inproc
worker pool at some point.

Uses the "dumb" API for now, which has a number of string copies.
IIRC the userspace/kernel copying is unavoidable, but maybe we
can look at using the smarter "zero copy" (in userspace) API
for messages. This can be significant if we're streaming large
volumes of data to-and-fro, e.g., the metrics data?

I do not think the metrics should be directly streamed to the
(primary / pilot) machine, but that is a separate issue for
a separate day.