Left outer join  (#1100)
* Optimizer changes to support left outer join

* Added functionality for naive iteration through hash table

Also temporary hot fix in binder, so I can test left outer join

* Various fixes in an attempt to get to left outer join to work

* Left outer join working with hacks

Fixed type declaration of OuterJoinConsumer function

* Various fixes

- Fix formatting
- Fix operator test
- Fix bug in how the WorkContext state is initialized

* Ugly solution to binder bug

Add a vector in addition to the table alias hash map to keep track of insertion order

* Made the naive hash table iterator more general

* Formatting

* Added pipeline state to arguements for outer_left_consumer function

* Proper implementation of hash table iterator

Previous iterator did not work for tables built in parallel.

Also revert adding pipeline state to arguement for OuterLeftConsumer. The idea
is that we need to pass in the pipeline state as an arguement to prevent race
conditions, but currently I'm not sure if the pipeline state is accessible
from the specific section of the code. So, I'm going to not pass the pipeline
state until recieving further feedback.

* Add left outer join test

Only one test has been added, definitely needs more coverage

* Formatting and removed unused functions

* Added pipeline state as third arguement for OuterLeftConsumer

* Add documentation for AdvancePipelineIter

* Increased maximum number of builtin functions

Number of builtins surpassed 256 so needed to upgrade from
a uint8_t to a uint16_t

* Documentation and Style changes

Use auto more to match existing style in hash_join_translator.cpp

Change outer_flag to left_outer_flag in order for it to be easier to
add future outer join functionality

Update comments

* Update with additional tests

* Renamed HashTableNaiveIterator to JoinHashTableIterator

* Added more tests

* Minor bug fix

* Refactored the way pipeline state variable is created in left outer join

* Formatting

* Revert pipeline state refactoring

* Various style improvements

- Moved JoinHashTableIterator into seperate cpp file

- Refactor AdvancePipelineIterator into SetSource
  and make it a bit more general

* Formatting

* Added test with two different types of tables

* Readded left outer join tests - merge shenanigans

* Refactored hash join translator code

Use joinConsumer even for regular hash joins, not solely for
left outer joins, this cuts down on code duplication in generated tpl

* Updated loops in pipeline with TODOs

* Formatting

* Fix typo in documentation

* too optimistic

* Only call joinConsumer in left outer join case

* Fix typo in tests

* Repeated table names caused errors with extended protocol

* Formatting

Co-authored-by: Matt Butrovich <mbutrovich@users.noreply.github.com>
Co-authored-by: Rohan Aggarwal <rohanagg@andrew.cmu.edu>