Hacky migration from libevent to libev
This commit migrates all libevent usages to libev. It is very
hacky and not clean.

It uses the libev C API and not the C++, using the C++ API is not
necessarily the right thing to do, but worth looking into.

There are mallocs thrown everywhere which is probably not great.
The various events and loops should probably be organized a lot
better cleaner.

Also timeouts are no longer handled properly. I haven't looked
into this much, but two ideas are
  1. See if libev has a way to handle timeouts for ev_io
  2. Create a composite event out of the ev_timer and ev_io