Remove EXPECT_DEATH tests from various optimizer tests.
These tests do not conform to googletest advisories with
regards to death tests, see section Death Test Styles.

More fundamentally, the tests do not seem useful.
The first thing all the tests do is call some function.
This function instantly blows up, only in debug mode,
because the lengths of two input vector arguments
are mismatched. Then the EXPECT_DEATH catches it.
It isn't like the assertion disappears from regular
code that is run in debug mode.

The fact that the optimizer functions wrap the
debug-only assert in ifndebug is another can of worms
that probably can be left to another day.