Add explicit dependencies to CMake. noisepage_objlib (spdlog) test_util libs (gtest gtest_main gmock gmock_main) benchmark_util libs (gtest gtest_main gmock gmock_main)
Dependencies are built in release because the debug versions
have unacceptable performance for coverage builds. Especially spdlog.
make appears to struggle to figure these dependencies out on its own.
spdlog is not added as a true dependency because the library name
changes between debug mode and release mode,
namely libspdlogd.a versus libspdlog.a.
Due to the different names, the build system gets confused.
Therefore the dependency is manually added.