Refactor the Messenger with Matt's suggestion.
Instead of sender and receiver callback IDs, which I got confused by,
use source and destination callback IDs. This is helpful when writing
callbacks that take in a ZmqResponse message, because now the source
is clearly the remote host and the destination is clearly the local host,
whereas previously sender is ambiguous since the receiver is now sending
a message to the original sender. That last sentence is itself confusing.