Fix having clauses in aggregates
When building the physical plan of an aggregate expression the group
by columns are hard coded into the output tuple 0, and the aggregate
values are hard coded into the output tuple 1.

When we were evaluating the output schema, we were saving the results
of all the columns in a SINGLE map of expression to value index. This
map was later used in ExpressionUtil::EvaluateExpression and
ExpressionUtil::ConverExprCVNodes to derive the Value Tuples of the
having clause.

The issue was that the above two functions take a list of maps, where
each map corresponds to a single tuple, and the index of the map in
the list is the tuple index. Therefore we needed two seperate maps,
one map for tuple 0 which contained the group by columns and one map
for tuple 1 which contained the aggregate values.

Fixes #1222