Add support for duplicate table alias names
SQL allows queries to have duplicate table aliases as long as they
are at different query depths. For example the following is valid SQL
    SELECT * FROM foo AS f WHERE f.a IN (SELECT a FROM foo AS f)
Additionally if a table is not given an alias then we assign it an
alias equal to it's name. Meaning that the following query will
result in "foo" being a duplicate table alias
    SELECT * FROM foo WHERE a IN (SELECT a FROM foo)
The optimizer assumes that all table aliases will be unique, so
these types of queries fail.

The solution is to change the type of the table alias in TableRef
from string to AliasType, and assign a unique serial number to all
TableRefs in a query.

The table name attribute of ColumnValueExpression actually
corresponds to the table alias. So one consequence of this change was
that we need to also update ColumnValueExpression to have an
AliasType attribute for it's table alias. We also need to make sure
that the serial number in the ColumnValueExpression's AliasType
matches the serial number in the TableRef's AliasType so they could
be matched up during optimization and query execution. This is
handled in the binder.