Hardcode LLVM_DIR to llvm@8 path on OSX.
brew is generally fragile as part of a build system, sometimes making
changes that require us to update the template VM images that we run on.
We will now move towards installing brew from scratch.
This is problematic in Jenkins because the setting of environment variables
happens before the packages.sh install step.
Since we don't update LLVM_DIR often, we might as well hardcode it.