Fix correlated subquery optimizer rules
Part of RewritePullFilterThroughAggregation and
DependentSingleJoinToInnerJoin rules adds a group by
column to the aggregation from one of the sides of the predicate.
When selecting a column we need to make sure that the column is at
the same depth as the aggregation or deeper. If the group by column
is part of the outer query and the aggregation part of the inner
query, then the aggregation has no way of accessing the column. A
higher value for depth means deeper in the querer and a lower value
for depth means more shallow in the query. The current code was
always setting the column from the left side of the predicate as the
group by column. This was sometimes causing the group by column to
be more shallow than the aggregation itself which caused the query
to error out.

Fixes #1404