Disable DecimalMultiplicationRegressioTest by default.
Because the test relies on an input randomly generated test file from the
decimal_multiplication_generator.py script, and there is no such file
checked into CI. So it is better to accurately report that the test is
for all intents and purposes disabled.

I wonder if we want to add random tests to our CI builds?
This is in some sense the perfect candidate: random test input, run test,
you can log the test input if test fails and discard it if test passes.
But I think CI is already flaky enough as it is, so maybe not now.