Add support for INSERT INTO SELECT statement. (#1454)
The implementation is modeled very closely after normal inserts, with two major differences.

In the optimizer the InsertSelect node has an output of the child Select node.
This output is then pushed down to the child select node in the InputColumnDeriver.

In the translator there are two differences.

a. At the start if it's an INSERT INTO SELECT it will prepare the child select translator. If it's a normal INSERT it will prepare the insert values.
b. When setting the projected row if it's an INSERT INTO SELECT it will set it from the child output, if it's a normal INSERT it will set it from the values.

Fixes #1158.