Add nasty boolean flag hack to track whether free(tvi) needs to be called in pipeline state teardown.
Nesting adds new complications.
The problem is that for a SeqScan, it might actually be nested inside another node like another SeqScan.
So PipelineState is actually not quite the right level for any variable that can appear nested,
because those will need to be setup/teardown'd per nested iteration instead of once per pipeline.

But we shove it into the pipeline state anyway and basically generate an if for whether it needs to be freed.

This raises the question of whether the other translators have similar issues.
If they do and our tests don't catch it, then we probably need better tests...