Checkpoint. Remove messenger_test.
messenger_test was introduced back when the messenger had no consumers.
messenger_test is now inherently racey in a way that can only be solved
with more shared memory hacks. Because the messenger automatically
retries sending messages until it receives an ack, the different forked
processes must be aware of the current state of the messenger in each
to avoid shutting down too early. Specifically,

1. replica1 -> primary: "blah"
2. primary -> replica1: "blah-ok"
3. replica1: received response, send ack for "blah-ok", shut down
But there is a race between whether the ack is received before the
primary retries with another "blah-ok" (which would crash it).
To solve this, you need to ack the ack or use more shared memory flags.
In my opinion, the test is becoming overcomplicated.

Rather than continue building that house of cards,
will test messenger through replication.