Collect metadata for compiled queries such as optimize time, etc.
This is gated behind the (currently uncontrollable) settings in
LLVMEngine::GetEngineSettings().

Add -print-llvm and -print-asm to TPL binary.
It is difficult to pick between just "noisepage-relevant" functions
and the kitchen sink of all functions right now. As a hack, n-squared
check before printing to see if the function was explicitly declared
in TPL.