Print REAL with fmt lib.
The old behavior of to_string is to always print to 6 decimal places.
Failure case: prints 0.333333333 as 0.333333.
The fmt lib will print as many decimal places as are available, but
will not be able to print trailing 0s.
Failure case: prints 0.333000 as 0.333.
If we want to print trailing 0s, we can use stringstream and std::fixed.
Failure case: prints 0.333 as 0.333000.

In general, there is no good solution here.
If precision matters, revive the decimal PR.